# comment, standard, optimize