--- a/test/index.js
+++ b/test/index.js
@@ -1,13 +1,11 @@
-
-var pull     = require('pull-stream')
+var pull = require('pull-stream')
 var pushable = require('../')
-var test     = require('tape')
+var test = require('tape')
 
 test('pushable', function (t) {
-
   var buf = pushable()
 
-  //should be a read function!
+  // should be a read function!
 
   t.equal('function', typeof buf)
   t.equal(2, buf.length)
@@ -21,11 +19,10 @@ test('pushable', function (t) {
     })
   )
 
-  //SOMETIMES YOU NEED PUSH!
+  // SOMETIMES YOU NEED PUSH!
 
   buf.push(1)
   buf.push(2)
   buf.push(3)
   buf.end()
-
 })