hey @dominictarr,

here's [your suggestion](https://github.com/Hypercubed/EventsSpeedTests/pull/2#issuecomment-228223419) implemented.

seems to help a bit, before:

```
Theoretical max x 20,120,961 ops/sec ±4.86% (10 runs sampled) *burn in*
MiniSignals x 15,381,183 ops/sec ±2.27% (10 runs sampled)
mini-pipe x 14,867,393 ops/sec ±6.16% (10 runs sampled)
EventEmitter x 13,120,457 ops/sec ±3.63% (10 runs sampled)
MicroSignals x 14,177,460 ops/sec ±15.69% (10 runs sampled)
EventEmitter3 x 12,476,861 ops/sec ±9.58% (10 runs sampled)
signal-lite x 12,015,678 ops/sec ±7.56% (10 runs sampled)
dripEmitter x 11,812,423 ops/sec ±11.84% (10 runs sampled)
EventEmitter2 x 14,187,397 ops/sec ±35.18% (10 runs sampled)
push-stream x 10,881,711 ops/sec ±6.01% (10 runs sampled)
ReactiveProperty x 8,956,755 ops/sec ±8.41% (10 runs sampled)
push-stream-patch x 9,345,109 ops/sec ±14.26% (10 runs sampled)
observ x 6,266,037 ops/sec ±11.30% (10 runs sampled)
RXJS x 3,349,851 ops/sec ±8.77% (10 runs sampled)
event-signal x 2,789,982 ops/sec ±5.97% (10 runs sampled)
barracks x 2,835,602 ops/sec ±14.77% (10 runs sampled)
d3-dispatch x 1,726,163 ops/sec ±7.91% (10 runs sampled)
pull-pushable x 1,587,159 ops/sec ±1.94% (10 runs sampled)
dripEmitterEnhanced x 853,680 ops/sec ±1.51% (10 runs sampled)
pull-notify x 437,368 ops/sec ±4.70% (10 runs sampled)
observable x 481,767 ops/sec ±19.53% (10 runs sampled)
signal-emitter x 342,021 ops/sec ±4.01% (10 runs sampled)
minivents x 279,941 ops/sec ±3.11% (10 runs sampled)
JS-Signals x 245,514 ops/sec ±14.47% (10 runs sampled)
namespace-emitter x 188,021 ops/sec ±2.07% (10 runs sampled)
```

after:

```
Theoretical max x 19,808,549 ops/sec ±3.05% (10 runs sampled) *burn in*
mini-pipe x 15,498,393 ops/sec ±7.05% (10 runs sampled)
MiniSignals x 13,925,170 ops/sec ±1.46% (10 runs sampled)
MicroSignals x 14,271,370 ops/sec ±7.38% (10 runs sampled)
signal-lite x 12,612,094 ops/sec ±8.09% (10 runs sampled)
push-stream x 11,406,447 ops/sec ±5.96% (10 runs sampled)
ReactiveProperty x 9,201,107 ops/sec ±11.20% (10 runs sampled)
push-stream-patch x 9,684,901 ops/sec ±33.43% (10 runs sampled)
EventEmitter x 7,338,195 ops/sec ±4.14% (10 runs sampled)
EventEmitter3 x 9,146,239 ops/sec ±35.05% (10 runs sampled)
dripEmitter x 8,782,591 ops/sec ±30.70% (10 runs sampled)
EventEmitter2 x 8,932,679 ops/sec ±38.77% (10 runs sampled)
observ x 4,372,794 ops/sec ±1.37% (10 runs sampled)
pull-pushable x 3,401,511 ops/sec ±9.42% (10 runs sampled)
event-signal x 3,043,959 ops/sec ±5.05% (10 runs sampled)
barracks x 2,569,803 ops/sec ±5.57% (10 runs sampled)
RXJS x 2,434,372 ops/sec ±6.69% (10 runs sampled)
d3-dispatch x 1,794,338 ops/sec ±4.04% (10 runs sampled)
dripEmitterEnhanced x 979,185 ops/sec ±7.60% (10 runs sampled)
observable x 396,305 ops/sec ±3.06% (10 runs sampled)
pull-notify x 351,958 ops/sec ±3.27% (10 runs sampled)
signal-emitter x 314,840 ops/sec ±4.41% (10 runs sampled)
minivents x 285,241 ops/sec ±5.31% (10 runs sampled)
JS-Signals x 204,678 ops/sec ±1.78% (10 runs sampled)
namespace-emitter x 182,191 ops/sec ±4.99% (10 runs sampled)
```

also standard-ized the overall module code and added comments.