@dominictarr 2016-06-26
merged into 2.0.1